pyproject.toml
[project]
name = "expresso"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "emotiefflib>=1.1.1",
    "facenet-pytorch>=2.5.3",
    "fastapi>=0.117.1",
    "jupyter>=1.1.1",
    "mediapipe>=0.10.21",
    "ml-dtypes>=0.4.1",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
    "websockets>=15.0.1",
]
